+++ rxjava-core/src/main/java/rx/Observable.java
@@ -49,87 +49,7 @@ import rx.observables.BlockingObservable;
 import rx.observables.ConnectableObservable;
 import rx.observables.GroupedObservable;
 import rx.observers.SafeSubscriber;
-import rx.operators.OnSubscribeFromIterable;
-import rx.operators.OnSubscribeRange;
-import rx.operators.OperationAll;
-import rx.operators.OperationAmb;
-import rx.operators.OperationAny;
-import rx.operators.OperationAsObservable;
-import rx.operators.OperationAverage;
-import rx.operators.OperationBuffer;
-import rx.operators.OperationCache;
-import rx.operators.OperationCombineLatest;
-import rx.operators.OperationConcat;
-import rx.operators.OperationDebounce;
-import rx.operators.OperationDefaultIfEmpty;
-import rx.operators.OperationDefer;
-import rx.operators.OperationDelay;
-import rx.operators.OperationDematerialize;
-import rx.operators.OperationDistinct;
-import rx.operators.OperationDistinctUntilChanged;
-import rx.operators.OperationElementAt;
-import rx.operators.OperationFinally;
-import rx.operators.OperationFlatMap;
-import rx.operators.OperationGroupByUntil;
-import rx.operators.OperationGroupJoin;
-import rx.operators.OperationInterval;
-import rx.operators.OperationJoin;
-import rx.operators.OperationJoinPatterns;
-import rx.operators.OperationMaterialize;
-import rx.operators.OperationMergeDelayError;
-import rx.operators.OperationMergeMaxConcurrent;
-import rx.operators.OperationMinMax;
-import rx.operators.OperationMulticast;
-import rx.operators.OperationOnErrorResumeNextViaObservable;
-import rx.operators.OperationOnErrorReturn;
-import rx.operators.OperationOnExceptionResumeNextViaObservable;
-import rx.operators.OperationParallelMerge;
-import rx.operators.OperationReplay;
-import rx.operators.OperationRetry;
-import rx.operators.OperationSample;
-import rx.operators.OperationSequenceEqual;
-import rx.operators.OperationSingle;
-import rx.operators.OperationSkip;
-import rx.operators.OperationSkipLast;
-import rx.operators.OperationSkipUntil;
-import rx.operators.OperationSkipWhile;
-import rx.operators.OperationSum;
-import rx.operators.OperationSwitch;
-import rx.operators.OperationSynchronize;
-import rx.operators.OperationTakeLast;
-import rx.operators.OperationTakeTimed;
-import rx.operators.OperationTakeUntil;
-import rx.operators.OperationTakeWhile;
-import rx.operators.OperationThrottleFirst;
-import rx.operators.OperationTimeInterval;
-import rx.operators.OperationTimer;
-import rx.operators.OperationToMap;
-import rx.operators.OperationToMultimap;
-import rx.operators.OperationToObservableFuture;
-import rx.operators.OperationUsing;
-import rx.operators.OperationWindow;
-import rx.operators.OperatorCast;
-import rx.operators.OperatorDoOnEach;
-import rx.operators.OperatorFilter;
-import rx.operators.OperatorGroupBy;
-import rx.operators.OperatorMap;
-import rx.operators.OperatorMerge;
-import rx.operators.OperatorObserveOn;
-import rx.operators.OperatorOnErrorResumeNextViaFunction;
-import rx.operators.OperatorOnErrorFlatMap;
-import rx.operators.OperatorParallel;
-import rx.operators.OperatorRepeat;
-import rx.operators.OperatorScan;
-import rx.operators.OperatorSubscribeOn;
-import rx.operators.OperatorTake;
-import rx.operators.OperatorTimeout;
-import rx.operators.OperatorTimeoutWithSelector;
-import rx.operators.OperatorTimestamp;
-import rx.operators.OperatorToObservableList;
-import rx.operators.OperatorToObservableSortedList;
-import rx.operators.OperatorUnsubscribeOn;
-import rx.operators.OperatorZip;
-import rx.operators.OperatorZipIterable;
+import rx.operators.*;
 import rx.plugins.RxJavaObservableExecutionHook;
 import rx.plugins.RxJavaPlugins;
 import rx.schedulers.Schedulers;
@@ -6274,7 +6194,7 @@ public class Observable<T> {
      * @see <a href="https://github.com/Netflix/RxJava/wiki/Filtering-Observables#wiki-skip">RxJava Wiki: skip()</a>
      */
     public final Observable<T> skip(int num) {
-        return create(OperationSkip.skip(this, num));
+        return lift(new OperatorSkip<T>(num));
     }
 
     /**
