+++ rxjava-core/src/main/java/rx/operators/OperationSkip.java
@@ -40,81 +40,6 @@ import rx.subscriptions.CompositeSubscription;
 public final class OperationSkip {
 
     /**
-     * Skips a specified number of contiguous values from the start of a Observable sequence and then returns the remaining values.
-     * 
-     * @param items
-     * @param num
-     * @return the observable sequence starting after a number of skipped values
-     * 
-     * @see <a href="http://msdn.microsoft.com/en-us/library/hh229847(v=vs.103).aspx">Observable.Skip(TSource) Method</a>
-     */
-    public static <T> OnSubscribeFunc<T> skip(final Observable<? extends T> items, final int num) {
-        // wrap in a Observable so that if a chain is built up, then asynchronously subscribed to twice we will have 2 instances of Take<T> rather than 1 handing both, which is not thread-safe.
-        return new OnSubscribeFunc<T>() {
-
-            @Override
-            public Subscription onSubscribe(Observer<? super T> observer) {
-                return new Skip<T>(items, num).onSubscribe(observer);
-            }
-
-        };
-    }
-
-    /**
-     * This class is NOT thread-safe if invoked and referenced multiple times. In other words, don't subscribe to it multiple times from different threads.
-     * <p>
-     * It IS thread-safe from within it while receiving onNext events from multiple threads.
-     * 
-     * @param <T>
-     */
-    private static class Skip<T> implements OnSubscribeFunc<T> {
-        private final int num;
-        private final Observable<? extends T> items;
-
-        Skip(final Observable<? extends T> items, final int num) {
-            this.num = num;
-            this.items = items;
-        }
-
-        public Subscription onSubscribe(Observer<? super T> observer) {
-            return items.subscribe(new ItemObserver(observer));
-        }
-
-        /**
-         * Used to subscribe to the 'items' Observable sequence and forward to the actualObserver up to 'num' count.
-         */
-        private class ItemObserver implements Observer<T> {
-
-            private AtomicInteger counter = new AtomicInteger();
-            private final Observer<? super T> observer;
-
-            public ItemObserver(Observer<? super T> observer) {
-                this.observer = observer;
-            }
-
-            @Override
-            public void onCompleted() {
-                observer.onCompleted();
-            }
-
-            @Override
-            public void onError(Throwable e) {
-                observer.onError(e);
-            }
-
-            @Override
-            public void onNext(T args) {
-                // skip them until we reach the 'num' value
-                if (counter.incrementAndGet() > num) {
-                    observer.onNext(args);
-                }
-            }
-
-        }
-
-    }
-
-    /**
      * Skip the items after subscription for the given duration.
      * 
      * @param <T>
