+++ rxjava-core/src/test/java/rx/operators/OperationSkipTest.java
@@ -17,7 +17,6 @@ package rx.operators;
 
 import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.*;
-import static rx.operators.OperationSkip.*;
 
 import java.util.concurrent.TimeUnit;
 
@@ -30,36 +29,6 @@ import rx.schedulers.TestScheduler;
 import rx.subjects.PublishSubject;
 
 public class OperationSkipTest {
-
-    @Test
-    public void testSkip1() {
-        Observable<String> w = Observable.from("one", "two", "three");
-        Observable<String> skip = Observable.create(skip(w, 2));
-
-        @SuppressWarnings("unchecked")
-        Observer<String> observer = mock(Observer.class);
-        skip.subscribe(observer);
-        verify(observer, never()).onNext("one");
-        verify(observer, never()).onNext("two");
-        verify(observer, times(1)).onNext("three");
-        verify(observer, never()).onError(any(Throwable.class));
-        verify(observer, times(1)).onCompleted();
-    }
-
-    @Test
-    public void testSkip2() {
-        Observable<String> w = Observable.from("one", "two", "three");
-        Observable<String> skip = Observable.create(skip(w, 1));
-
-        @SuppressWarnings("unchecked")
-        Observer<String> observer = mock(Observer.class);
-        skip.subscribe(observer);
-        verify(observer, never()).onNext("one");
-        verify(observer, times(1)).onNext("two");
-        verify(observer, times(1)).onNext("three");
-        verify(observer, never()).onError(any(Throwable.class));
-        verify(observer, times(1)).onCompleted();
-    }
 
     @Test
     public void testSkipTimed() {
